requirements.txt
streamlit
pandas
google-genai
requests
beautifulsoup4